[LEVWEB-792] Use Oracle as the source database
This is probably a more typical use-case and allows us to side-step what
appears to be a bug in the MySQL support.

The downside is that the Oracle image is ugly and listens on it's port
before it is actually ready to accept connections, requiring us to sleep
for a certain time to allow for this.