[LEVWEB-1138] Standard logging plus timestamps
Reinstates the original config file for controlling logging but makes
some changes to achieve the following:

1. Sets certain log levels based on environment variables
2. Adds extra configuration to ensure SQL statements are not leaked
   accidentally
3. Adds timestamps to the logs
4. Adds the log level to the logs

A new environment variable `SYMDS_LOG_LEVEL` has been created to cover
SymmetricDS as opposed to the libraries that it calls. It defaults to
INFO making this image a bit more chatty than it has been. (Though it
has been too quiet since we last changed the logging.)